requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
Click==7.0
Flask==1.0.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.16.0
Pillow==5.4.1
protobuf==3.6.1
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
